astro: refactor info
