+++ packages/astro-example-floors/src/utils/info.ts
@@ -1,0 +1,22 @@
+import { z } from 'zod'
+
+export const infoSchema = z.discriminatedUnion('tag', [
+  z.object({
+    tag: z.literal('shop.cafe'),
+  }),
+  z.object({
+    tag: z.literal('shop.misc'),
+  }),
+  z.object({
+    tag: z.literal('shop.restaurant'),
+  }),
+  z.object({
+    tag: z.literal('facility.elevator'),
+  }),
+  z.object({
+    tag: z.literal('facility.escalator'),
+  }),
+  z.object({
+    tag: z.literal('facility.toilet'),
+  }),
+])
