+++ packages/astro-example-floors/src/utils/schema.ts
@@ -58,27 +58,6 @@ export const poiSchema = z.object({
   x: poixSchema,
 })
 
-export const infoSchema = z.discriminatedUnion('tag', [
-  z.object({
-    tag: z.literal('shop.cafe'),
-  }),
-  z.object({
-    tag: z.literal('shop.misc'),
-  }),
-  z.object({
-    tag: z.literal('shop.restaurant'),
-  }),
-  z.object({
-    tag: z.literal('facility.elevator'),
-  }),
-  z.object({
-    tag: z.literal('facility.escalator'),
-  }),
-  z.object({
-    tag: z.literal('facility.toilet'),
-  }),
-])
-
 // searchAddressesSchema
 // searchNamesSchema
